requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
django-extensions==3.1.3
ipdb==0.13.9
python-decouple==3.4